requirements.txt
# Pacotes originais
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
openpyxl>=3.0.0
numpy>=1.21.0

# Flask e extensões
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
python-dotenv==1.0.0
email-validator==2.1.0
SQLAlchemy==2.0.36
psycopg2-binary>=2.9.9
requests>=2.28.0
python-dateutil>=2.8.2
